redoing signals to match specefications better
